Pull nested types out. Document them better.
