+++ datastore/src/main/java/io/spine/server/storage/datastore/query/ConvertAsIs.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2021, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.storage.datastore.query;
+
+import com.google.cloud.datastore.Entity;
+import com.google.protobuf.FieldMask;
+import com.google.protobuf.Message;
+import io.spine.type.TypeUrl;
+
+import java.util.stream.Stream;
+
+/**
+ * Converts the original {@code Entity} instances into the records of {@code <R>}
+ * and applies the field mask to each of them.
+ *
+ * <p>Performs no intermediate filtering.
+ *
+ * @param <R>
+ *         the type of the records to convert each {@code Entity} into
+ */
+final class ConvertAsIs<R extends Message> extends ToRecords<R> {
+
+    /**
+     * Creates a new instance of this function.
+     */
+    ConvertAsIs(TypeUrl type, FieldMask mask) {
+        super(type, mask);
+    }
+
+    @Override
+    protected Stream<Entity> filter(Stream<Entity> entities) {
+        return entities;
+    }
+}
